rewrite indexer to be a listener of eventBus

+++ rpc/core/types/responses.go
@@ -107,12 +107,12 @@ type ResultBroadcastTxCommit struct {
 	CheckTx   abci.Result `json:"check_tx"`
 	DeliverTx abci.Result `json:"deliver_tx"`
 	Hash      data.Bytes  `json:"hash"`
-	Height    int         `json:"height"`
+	Height    uint64      `json:"height"`
 }
 
 type ResultTx struct {
-	Height   int           `json:"height"`
-	Index    int           `json:"index"`
+	Height   uint64        `json:"height"`
+	Index    uint32        `json:"index"`
 	TxResult abci.Result   `json:"tx_result"`
 	Tx       types.Tx      `json:"tx"`
 	Proof    types.TxProof `json:"proof,omitempty"`
